requirements.txt
Django==2.0.7
numpy==1.14.5
pymongo==3.7.1
djangorestframework==3.8.2
django-oauth-toolkit==1.2.0